feat: 작품 관련 model 구현
